Changed form name & some validation changes
